feat: move util functions in main to root utils file
